Replace " with &quot; entity in ckeditors alt text. Also add .arcconfig
Summary:
A few times now translation process has died because of articles
created with this bug in it.  The unescaped quotes before would
end the alt tag and the rest of the alt message would become attributes
on the element. If any of that text was a pure number, then it would
fail to parse as valid html during fake_translation

Test Plan:
./dev/builder/build.sh
From any article, in edit mode add an image that has a quote in the alt text
Check it out in the inspector, to see the alt text showing up correctly.
Publish
Change the src in the inspector, make sure the alt shows correctly

Reviewers: tom

Differential Revision: http://phabricator.khanacademy.org/D5278

Conflicts:
	plugins/image2/plugin.js