Add new "drawing" plugin to CKeditor.
Summary: This uses the Method-draw page that we load on KA.org. This creates a new plugin, and dialog, for creating and editing SVG created using Method-Draw.

Test Plan: Open the editor, click the "Add Drawing" button, the dialog should open and you should be able to create (and insert) a new drawing. You can edit the drawing by double-clicking it.

Reviewers: pamela, meredith

Reviewed By: meredith

Subscribers: colin, #cs

Differential Revision: https://phabricator.khanacademy.org/D20046