Fix a number of issues relating to the drawing plugin.
Summary:
1) When you have a data: url image inside an SVG that's inside an img element, the data: url image won't render. Thus we need to use an iframe just for safari. Other browsers just use a normal img element.
2) Provide an overlay that makes it obvious that you need to click (or double-click, in the case of IE) the image in order to edit it.
3) Fix the SVG data URL so that the charset is specified correctly (the old format errored out in IE).
4) Marked up strings for translation (hopefully will work!).

I also added copious comments to help clarify things a bit.

The only remaining bug, that I eventually gave up on, is that in IE 9 the "Double-click to edit" overlay doesn't pop up. But it still works - the user can double-click just fine. IE 11 works as you would expect it to.

Test Plan: Tested and verified it works in Chrome, Firefox, Safari, IE 9, and IE 11.

Reviewers: meredith

Reviewed By: meredith

Differential Revision: https://phabricator.khanacademy.org/D21450